requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
base32hex==1.0.2
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cx-Oracle==8.3.0
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.2.0
django-rest-auth==0.9.5
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
Jinja2==3.1.2
Pillow==9.5.0
PyJWT==2.6.0
PyPDF2==3.0.1
python-decouple==3.8
weasyprint==61.0
psycopg2==2.9.9
pandas==2.0.1
paramiko==3.1.0